Change the type of the result of the screen index calculation to size_t

+++ include/gos/screen.h
@@ -1,13 +1,15 @@
 #ifndef GOS_SCREEN_H_
 #define GOS_SCREEN_H_
 
+#include <stddef.h>
+
 #include <gos/types.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-int gos_screen_index(gos_screen* screen, int x, int y);
+size_t gos_screen_index(gos_screen* screen, int x, int y);
 
 #ifdef __cplusplus
 }
